Updated modules/modis-prep.js (added 1 line, deleted 1 line).

+++ modules/modis-prep.js
@@ -2,7 +2,7 @@
 
 
 exports.rescale = function(img) {
-  return img.multiply(0.0001).copyProperties(img);
+  return img.multiply(0.0001).float().copyProperties(img);
 };
